Use `.write` instead of `printfl!` to write `""`
Now that `printfl!` calls `write_all`, rather than `write`, this no
longer works. The `write_all` implementation provided by the `Write`
trait writes until the buffer of bytes left to write is empty, so `write`
is never if an empty string is passed, and thus a `Pipe` doesn't send a
message with `done`.

There are probably better ways to deal with the APIs involved here, but
this at least fixes the behavior.